Bug in restart mode fixed for Catlearn_NEB and Catlearn_AutoNEB.

+++ catlearn/optimize/catlearn_autoneb_optimizer.py
@@ -82,22 +82,6 @@ class CatLearnAutoNEB(object):
         self.initial_image = read(start, '-1')
         self.final_image = read(end, '-1')
 
-        # A and B) Restart mode.
-
-        # Read the previously evaluated structures and append them to the
-        # training list
-        if restart is True:
-            restart_filename = 'evaluated_structures.traj'
-            if not os.path.isfile(restart_filename):
-                warning_restart_neb()
-            if os.path.isfile(restart_filename):
-                evaluated_images = read(restart_filename, ':')
-                is_endpoint = evaluated_images + is_endpoint
-
-        # Write previous evaluated images in evaluations:
-        write('./evaluated_structures.traj',
-              is_endpoint + fs_endpoint)
-
         # Check the magnetic moments of the initial and final states:
         self.magmom_is = is_endpoint[-1].get_initial_magnetic_moments()
         self.magmom_fs = fs_endpoint[-1].get_initial_magnetic_moments()
@@ -115,12 +99,21 @@ class CatLearnAutoNEB(object):
         self.scale_targets = np.min([energy_is, energy_fs])
 
         # Convert atoms information into data to feed the ML process.
-        if os.path.exists('./tmp.traj'):
-                os.remove('./tmp.traj')
-        merged_trajectory = is_endpoint + fs_endpoint
-        write('tmp.traj', merged_trajectory)
 
-        trj = ase_traj_to_catlearn(traj_file='tmp.traj')
+        # Include Restart mode.
+
+        if restart is not True:
+            if os.path.exists('./tmp.traj'):
+                    os.remove('./tmp.traj')
+            merged_trajectory = is_endpoint + fs_endpoint
+            write('tmp.traj', merged_trajectory)
+            trj = ase_traj_to_catlearn(traj_file='tmp.traj')
+            os.remove('./tmp.traj')
+            write('./evaluated_structures.traj', is_endpoint + fs_endpoint)
+
+        if restart is True:
+            trj = ase_traj_to_catlearn(traj_file='./evaluated_structures.traj')
+
         self.list_train, self.list_targets, self.list_gradients, trj_images,\
             self.constraints, self.num_atoms = [trj['list_train'],
                                                 trj['list_targets'],
@@ -128,7 +121,6 @@ class CatLearnAutoNEB(object):
                                                 trj['images'],
                                                 trj['constraints'],
                                                 trj['num_atoms']]
-        os.remove('./tmp.traj')
         self.ase_ini = trj_images[0]
         self.num_atoms = len(self.ase_ini)
         if len(self.constraints) < 0:
